test again to be sure

+++ jax_triton/pallas/ops/attention.py
@@ -67,7 +67,7 @@ def mha_forward_kernel(
     if has_bias:
         bias = pl.load(bias_ref, (pl.dslice(start_q * block_q, block_q), pl.dslice(start_k * block_k, block_k))).astype(jnp.float32)
         qk += bias
-        qk = qk.astype(jnp.float32)
+        # qk = qk.astype(jnp.float32)
 
     if causal:
       span_q = start_q * block_q + jnp.arange(block_q)
@@ -300,7 +300,7 @@ def mha_backward_kernel(
       if has_bias:
           bias = pl.load(bias_ref, (pl.dslice(start_q * block_q, block_q), pl.dslice(start_k * block_k, block_k))).astype(jnp.float32)
           qk += bias
-          qk = qk.astype(jnp.float32)
+          # qk = qk.astype(jnp.float32)
 
 
       if causal:
